+++ HSTracker/Logging/Parsers/TagChangeActions.swift
@@ -234,6 +234,23 @@ struct TagChangeActions {
             // Card was created by Suspicious Alchemist/Usher/Pirate
             return
         }
+        
+        let currentBlock = AppDelegate.instance().coreManager.logReaderManager.powerGameStateParser.currentBlock
+        
+        // prevents nightmare fuel and dejavu leaking the card
+        if (currentBlock?.cardId == CardIds.Collectible.Rogue.NightmareFuel ||
+            (currentBlock?.parent?.cardId == CardIds.Collectible.Rogue.NightmareFuel &&
+             entity.cardId == CardIds.NonCollectible.Neutral.TreacherousTormentor_DarkGiftToken) ||
+            currentBlock?.cardId == CardIds.Collectible.Rogue.DejaVu) &&
+            entity.isControlled(by: eventHandler.player.id) {
+            targetEntity.cardId = ""
+            targetEntity.info.hidden = true
+            if !entity.info.latestCardId.isEmpty {
+                eventHandler.opponent.predictUniqueCardInDeck(cardId: entity.info.latestCardId, isCreated: false)
+                AppDelegate.instance().coreManager.game.updateTrackers()
+            }
+            return
+        }
 
         if targetEntity.cardId == "" {
             targetEntity.cardId = entity.info.latestCardId